# libs/shared/api/lotchen-client-api/src/lib/model/filter-base.interface.ts
/**
 * Talisoft.Lotchen.API
 * The Lotchen API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FilterBase { 
    /**
     * Global Text search
     */
    globalText: string;
}