# libs/shared/api/lotchen-client-api/src/lib/model/find-all-territories-query-response.interface.ts
/**
 * Talisoft.Lotchen.API
 * The Lotchen API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FindAllTerritoriesQueryResponse { 
    id: string;
    name: string;
    /**
     * Date of creation
     */
    createdAt: string;
    /**
     * Person who created the entry
     */
    createdBy: string;
}