# libs/shared/api/openapi-documents/lotchen-api-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_createUser_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCommand"
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_allUsers_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllUserQuery"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findUserById_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindUserByIdQueryResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/assign-permissions": {
      "put": {
        "operationId": "UsersController_assignPermissionsToUser_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionsCommandRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/assign-roles": {
      "put": {
        "operationId": "UsersController_assignRolesToUser_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesCommandRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/roles": {
      "get": {
        "operationId": "UsersController_updateUserRoles_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/permissions": {
      "get": {
        "operationId": "UsersController_updateUserPermissions_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUserPermissionsQueryResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/search": {
      "post": {
        "operationId": "UsersController_searchUserPaginate_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginateAllUsersCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "title": "PaginationResponseOfPaginateAllUsersCommandDto",
                      "$ref": "#/components/schemas/Pagination"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PaginateAllUsersCommandDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/territories": {
      "post": {
        "operationId": "TerritoriesController_createTerritory_v1",
        "parameters": [
          {
            "name": "x-tenant-fqdn",
            "in": "header",
            "description": "The Tenant Fqdn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTerritoryCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Territories"
        ]
      },
      "get": {
        "operationId": "TerritoriesController_allTerritories_v1",
        "parameters": [
          {
            "name": "x-tenant-fqdn",
            "in": "header",
            "description": "The Tenant Fqdn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindAllTerritoriesQueryResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Territories"
        ]
      }
    },
...
    "/api/v1/teams": {
      "post": {
        "operationId": "TeamsController_createTeam_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamCommand"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTeamCommandResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_findAllTeams_v1",
        "parameters": [
          {
            "name": "x-tenant-fqn",
            "in": "header",
            "description": "The Tenant Fqn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindAllTeamsQueryResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/organizations": {
      "get": {
        "operationId": "OrganizationsController_findAllOrganizations_v1",
        "parameters": [
          {
            "name": "x-tenant-fqdn",
            "in": "header",
            "description": "The Tenant Fqdn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindAllOrganizationQueryResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_createOrganization_v1",
        "parameters": [
          {
            "name": "x-tenant-fqdn",
            "in": "header",
            "description": "The Tenant Fqdn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationCommand"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    }
  },
  "info": {
    "title": "Talisoft.Lotchen.API",
    "description": "The Lotchen API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "confirmPassword"
        ]
      },
      "GetAllUserQuery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "description": "User email "
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "SchemaObjectId": {
        "type": "object",
        "properties": {}
      },
      "FindUserByIdQueryRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "FindUserByIdQueryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SchemaObjectId"
          },
          "email": {
            "type": "string"
          },
          "isVerified": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          },
          "isSuperAdmin": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roles": {
            "$ref": "#/components/schemas/FindUserByIdQueryRoleDto"
          }
        },
        "required": [
          "id",
          "email",
          "isVerified",
          "isLocked",
          "isSuperAdmin",
          "permissions",
          "roles"
        ]
      },
      "AssignPermissionsCommandRequest": {
        "type": "object",
        "properties": {
          "permissions": {
            "description": "List of permissions keys",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "AssignRolesCommandRequest": {
        "type": "object",
        "properties": {
          "roles": {
            "description": "List of permissions keys",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roles"
        ]
      },
      "GetUserPermissionsQueryResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "number",
            "description": ""
          },
          "pageIndex": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "totalElements",
          "pageIndex",
          "pageSize",
          "totalPages",
          "data"
        ]
      },
      "PaginateAllUsersTeamDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "PaginateAllUsersCommandDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isVerified": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaginateAllUsersTeamDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "isVerified",
          "isActive",
          "roles",
          "permissions",
          "teams",
          "createdAt",
          "updatedAt"
        ]
      },
      "FilterBase": {
        "type": "object",
        "properties": {
          "globalText": {
            "type": "string",
            "description": "Global Text search"
          }
        },
        "required": [
          "globalText"
        ]
      },
      "PaginateAllUsersCommand": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/FilterBase"
          },
          "sort": {
            "type": "object"
          },
          "pageIndex": {
            "type": "number",
            "description": "Page Index",
            "example": 0,
            "minimum": 0
          },
          "pageSize": {
            "type": "number",
            "maximum": 30,
            "minimum": 5,
            "example": 5
          }
        },
        "required": [
          "filters",
          "pageIndex",
          "pageSize"
        ]
      },
...
      "CreateTerritoryCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the territory",
            "example": "Ivory Coast"
          },
          "description": {
            "type": "string"
          },
          "childrenIds": {
            "description": "Territory children",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "parentId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "FindAllTerritoriesQueryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date of creation"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Latest updated At"
          },
          "createdBy": {
            "type": "string",
            "description": "Person who created the entry"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
...
      "CreateTeamCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the team"
          },
          "description": {
            "type": "string"
          },
          "managerId": {
            "type": "string"
          },
          "memberIds": {
            "description": "User Ids to link into the team",
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateTeamCommandResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "FindAllTeamsQueryUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "FindAllTeamsQueryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "manager": {
            "$ref": "#/components/schemas/FindAllTeamsQueryUserDto"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FindAllTeamsQueryUserDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "manager",
          "members"
        ]
      },
      "FindAllOrganizationQueryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrganizationCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the organization (Subsidiary)"
          },
          "description": {
            "type": "string",
            "description": "Description of the organization, scope and country where is located"
          }
        },
        "required": [
          "name",
          "description"
        ]
      }
    }
  }
}